requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
authclient==1.0
blank==0.1.0
boto3==1.18.23
botocore==1.21.23
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
cryptocode==0.1
Django==3.2.5
django-crontab==0.7.1
django-filter==2.4.0
djangorestframework==3.12.4
docopt==0.6.2
ecdsa==0.17.0
enum-compat==0.0.3
future==0.18.2
idna==3.2
jmespath==0.10.0
jose==1.0.0
multidict==5.1.0
oauthlib==3.1.1
packaging==21.0
pyasn1==0.4.8
pycparser==2.18
pycryptodomex==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
python-jose==3.3.0
pytz==2021.1
qb==0.0
rauth==0.7.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
selenium==3.141.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.2.0
yarl==1.6.3
